Prefer Neovim as main editor
Relegate Vim to being a fallback if Neovim is unavailable. Add some
useful plugins (frozen in time) for the sake of editing but eschew
anything that's IDE-like which belongs with Neovim.

Neovim's configuration has been converted to Lua for the sake of
consistency with a lot of stuff that's coming up in that ecosystem.