Add various popup workflows to tmux
I'm going down the route of preferring one session per-project.
Introduce `att` as a simple script to find eligible projects to start a
new tmux session with (or re-use one if it exists).

Additionally, add some quality of life improvements in the form of a
scratch session (so I stop roaming outside of the tmux client) and a URL
picker. Allow for other super secret ones in a `tmux.local.conf` file...